Make sure type is preserved after division

+++ image_similarity_measures/quality_metrics.py
@@ -27,7 +27,7 @@ def rmse(org_img: np.ndarray, pred_img: np.ndarray, max_p=4095) -> float:
 
     rmse_bands = []
     for i in range(org_img.shape[2]):
-        m = np.mean(np.square((org_img[:, :, i] - pred_img[:, :, i]) / max_p))
+        m = np.mean(np.square(((org_img[:, :, i] - pred_img[:, :, i]) / max_p).astype(org_img.dtype)))
         s = np.sqrt(m)
         rmse_bands.append(s)
